Fix warnings from static analysis tools
1.CHECKED_RETURN:
  a. NativeNfcTag.cpp: nativeNfcTag_doWrite()
                       nativeNfcTag_makeMifareNdefFormat ()
                       nativeNfcTag_doNdefFormat()
  Added checks for return values of APIs

2.INVALID_ASSIGNMENT_WITH_UNSIGNED_VALUES:
  a. phNxpExtns_MifareStd.cpp: phFriNfc_ExtnsTransceive()

+++ nci/jni/NativeNfcTag.cpp
@@ -1720,7 +1720,7 @@ static jbyteArray nativeNfcTag_doTransceive(JNIEnv* e, jobject o,
         DLOG_IF(INFO, nfc_debug_enabled)
             << StringPrintf("%s: reconnect finish", __func__);
       } else if (sCurrentConnectedTargetProtocol == NFC_PROTOCOL_MIFARE) {
-        uint32_t transDataLen = sRxDataBuffer.size();
+        uint32_t transDataLen = static_cast<uint32_t>(sRxDataBuffer.size());
         uint8_t* transData = (uint8_t*)sRxDataBuffer.data();
         bool doReconnect = false;
